refactor(io): split daft-mcap into scan and reader modules
lib.rs keeps the scan surface (options, summary reading, stream_mcap and
the prefetch pipeline); read.rs holds NativeMcapReader with the indexed
chunk traversal and linear fallback. Tests move to the bottom of the
module they exercise, with shared fixtures in a cfg(test) test_utils
module. Pure code movement, no behavior change.
